+++ src/components/validation/validator.js
@@ -498,11 +498,6 @@ Validator.prototype = {
       if (!($(`.icon-${type}`, iconContainer).length) && !parentContainer.hasClass('module-tabs')) {
         iconContainer.addClass(`is-${type}`).append(errorIcon);
       }
-
-      const tabsAPI = parentContainer.data('tabs');
-      if (tabsAPI) {
-        tabsAPI.sizeBar();
-      }
     } else {
       // Remove icon
       iconContainer = iconContainer.add(menuitem);